Cleaned up action extension, fixed iPad popup bug
The Safari Extension popup on iPad now appears at a consistent size.
Instead of creating a new QR code when the user switches from light to dark mode, the original QR code now has its colors inverted.
Added an icon and name to the Action Extension.
Action Extension now accepts URLs in the form of Strings, as well as URLs in the form of URLs.
Image save button changed to an SF Symbol.
Added information about the Action Extension to the main app's tip page.